requirements.txt
peewee
flask
python-magic
